image_ai_utils/requirements.txt
pillow
python-dotenv
httpx
pydantic
websocket-client